+++ src/main/java/coffee/synyx/frontpage/installer/PluginInstallationException.java
@@ -1,0 +1,11 @@
+package coffee.synyx.frontpage.installer;
+
+/**
+ * @author Tobias Schneider
+ */
+public class PluginInstallationException extends RuntimeException {
+
+    PluginInstallationException(String message) {
+        super(message);
+    }
+}